feat: split handler
